keep canvas synchronized
desynchronized is for things like pen lines, where we don't need to refresh the whole screen.

Since we're redrawing the whole screen, it just means we get crazy page tearing.

In the future we probably want to do some better draw scheduling with requestAnimationFrame, and estimates for how long drawing will take, so that we can finish drawing close the the frame deadline.

Or even just use requestAnimationFrame in the draw Debouncer?